Added Parallax script and made test backdrops
This script takes in three inputs, the main camera, the starting offset, and the amount of parallax scrolling each plane will move relative to the player's movement. The planes on which the backdrop images will be displayed need a script component and also a sprite render. Added a few test backdrops to test scene to demonstrate it functioning.