MS1:AttackingMechanics
I added a sword mechanic for the player and changed the shooting mechanics so that you can aim and an arrow will show up to where you are pointing. You can also cancel the aim. The objects/scripts were added to the character prefab. I copied and removed ThrowManager from the character and added ThrowTest because I didn't want to change ThrowManager. My Scripts are in Player Scripts named ThrowTest and SwordSwing. My objects are named ArrowRotate and SwordRotate. I made a object called SideSwitch that checks where the mouse is when you swing the sword so it will be on the left or right of the character. I also added sprites for the arrow and the Bone Sword