Created the sporeInventory, Life System, Collectables Spores and coins and the UI based of this
I created a script called SporeItem which is the main way to differentiate the spores there is the spore prefab and all informations about and it's easy to change things about this.
The SporeSelect which is the sporeInventory, it has a list of current spores, we can add spores to it like the collectable I did and we can switch through the different spores with 1 and 2 now, there is also animations to switch between them but nothing when obtained.
There is a Collectable script that is used for collectables, it's easy to understand what it does, there is only two collectables, there spores and coins, UI added for the coins too.
There is a life system added which is really easy to understand too.