World Manager System
This branch is dedicated to figuring out how to organize "rooms" in the world, which are individual organizations of environments and their collisions, door triggers, enemies, backgrounds, and so on that are specific to that room. The world manager object with the room management script handles the loading and unloading of rooms within a single scene. This script handles activating the appropriate "room" when the player enters a trigger. Then the current room is deactivated. Triggers need to be placed at the entrances and exits. In the inspector, you can add room prefabs to the "rooms" array. Rooms[0] will hold the starting area. There is a "door" trigger with a script that simply checks collision with the player. A major issue I wasn't able to solve yet was adjusting where the player spawns relative to the prefab room so you don't fall through the floor.