MS3:FrogEnemy/HealthDrop
I created a frog like enemy that jumps and follows the enemy when in range. If you enter a second range the frog will send out its tongue. I also gave all enemies a 1 in 5 chance to drop a health collectable. The only issue I have is getting the tongue to unroll from the mouth and retracting.