I have changed a couple scripts and prefabs so the player can actually play the game.
Right now the spikes on the wall jump level have been turned off but you can turn them back on. I recommend getting rid of the spike right next to the exit. It just seems a little bit too hard but if you want to keep it that way its cool. The checkpoints were not on 0 z axis so it was messing with game when you respawn so i fixed that.